+++ src/muml/video.nim
@@ -7,16 +7,16 @@ proc getVideo* (muml: mumlNode): mumlObject =
     of "path":
       result.path = val.getStr.removeDoubleQuotation
     of "frame":
-      result.frame.start = val["start"].getFloat
-      result.frame.`end` = val["end"].getFloat
+      result.frame.start = val["start"].getInt
+      result.frame.`end` = val["end"].getInt
     of "layer":
       result.layer = val.getInt
     of "video":
       for key2, val2 in val.pairs:
         case key2:
         of "frame":
-          result.video.frame.start = val2["start"].getFloat
-          result.video.frame.`end` = val2["end"].getFloat
+          result.video.frame.start = val2["start"].getInt
+          result.video.frame.`end` = val2["end"].getInt
         of "position": result.video.position = val2.getPosition
         of "scale": result.video.scale = val2.getScale
         of "rotate": result.video.rotate = val2.getNumberValue
